requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
tensorflow==2.15.0
keras==2.15.0
Pillow==10.1.0
werkzeug==3.0.1
gunicorn==21.2.0